Skip 'promise > meow async' on Windows
This test checks whether the file `meow.cat` is executable. On Windows,
it determines this by checking whether the file extension is in the
PathExt environment variable. `cat` is not typically in PathExt, so
this test will always fail on Windows.

The functionality being tested is already covered by similar
windows-only tests (`windows > pathExt env > meow async` in particular),
so it should be safe to ignore without reducing coverage.

Closes #19